refactor: move WorkflowActionGroupFetch to RunnerBarCore/Runner (#923)
- Delete Sources/RunnerBar/Runner/WorkflowActionGroupFetch.swift
- Add Sources/RunnerBarCore/Runner/WorkflowActionGroupFetch.swift with public API
- Remove import RunnerBarCore (can't import yourself)
- makeActiveJob wrapper calls internal RunnerBarCore.makeActiveJob directly
- No import changes needed in callers (covered by @_exported in Exports.swift)
